% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig.R
\name{fig}
\alias{fig}
\title{Return figures used for pmder data visualization}
\usage{
fig(fig_name, color_style = "plasma")
}
\arguments{
\item{fig_name}{string, Name of the figure.}

\item{color_style}{string, color to be passed to scale_colour_viridis}
}
\description{
Return figures used for pmder data visualization
}
\examples{
fig("prob_det_energy") will return the efficiency as a function of the energy
fig("") will return a list of potential figures
}
