% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrated_counts_factor.R
\name{integrated_counts_factor}
\alias{integrated_counts_factor}
\title{Multiplicative factor for net counts}
\usage{
integrated_counts_factor(x1, x2)
}
\arguments{
\item{x1}{Final position in cm}

\item{x2}{Initial position in cm}
}
\value{
Vector containing multiplicative factors for each set of positions
}
\description{
Multiplicative factor to get from net count rate to net counts. Called by `predict_count_rate`.
}
