% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intrinsic_efficiency.R
\name{intrinsic_efficiency}
\alias{intrinsic_efficiency}
\title{Calculate intrinsic efficiency of data}
\usage{
intrinsic_efficiency(intrinsic_data)
}
\arguments{
\item{intrinsic_data}{a}
}
\value{
a
}
\description{
Calculate intrinsic efficiency of data
}
