% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcnp_input_creation.R
\name{mcnp_input_creation}
\alias{mcnp_input_creation}
\title{Create MCNP inputs for portal monitors}
\usage{
mcnp_input_creation(energy_list_kev, truck_position_list_cm, content_list)
}
\arguments{
\item{energy_list_kev}{List of photon source energies in kiloelectron-volts}

\item{truck_position_list_cm}{List of truck positions in centimeters}

\item{content_list}{`c("f", "m")` Contents of interior}
}
\description{
Given a list of source and truck configurations, the MCNP decks will be written along with a bash script to execute the runs.
}
