% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_count_rate.R
\name{predict_count_rate}
\alias{predict_count_rate}
\title{Title}
\usage{
predict_count_rate(
  input_data,
  photon_energy_keV,
  yield,
  activity_Bq = 1,
  method = "earth",
  earth_data = earth_model,
  cpm = FALSE,
  net_counts_rolling = FALSE,
  net_counts_cumulative = FALSE,
  count_rate_derivative = FALSE,
  collection_time_s = NULL,
  speed_cm_s = NULL
)
}
\arguments{
\item{speed_cm_s}{}
}
\description{
Title
}
